analysis_engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
prometheus_client==0.20.0
pydantic==2.9.2
upstash-redis==1.0.0

opencv-python-headless==4.10.0.84
numpy==2.0.2
Pillow==10.4.0
imagehash==4.3.1

# Optional accelerators / detectors (best-effort imports in code)
torch==2.5.1
ultralytics==8.3.38

